Refactor to prevent cyclic dependency (happened when importing the robusta package on the relay) Fix jobs cache initialization - Init only once, even on clusters with no jobs Fix silence url building - Build the url without making an http call (this failed for user with outgoing firewall)
